+++ packaging/homebrew/keenetic-manager.rb
@@ -1,8 +1,8 @@
 class KeeneticManager < Formula
   desc "Inspect Keenetic routers and manage client policies"
   homepage "https://github.com/osteotek/keenetic-manager"
-  url "https://github.com/osteotek/keenetic-manager/releases/download/v1.2.0/keenetic-manager-1.2.0.tar.gz"
-  sha256 "8b35bfb959b4e2773b0b048beff248cba1a6a1ca655cf6ea4f3063e97da6ed37"
+  url "https://github.com/osteotek/keenetic-manager/releases/download/v1.3.0/keenetic-manager-1.3.0.tar.gz"
+  sha256 "995bd559270737b150d6703079817a74a5e8e8a2cbc092511c6eca3db15b714a"
   license "MIT"
 
   depends_on "bash"
@@ -15,6 +15,6 @@ class KeeneticManager < Formula
   end
 
   test do
-    assert_match "keenetic 1.2.0", shell_output("#{bin}/keenetic --version")
+    assert_match "keenetic 1.3.0", shell_output("#{bin}/keenetic --version")
   end
 end